requirements.txt
asgiref==3.5.2
boto3==1.24.67
botocore==1.27.67
certifi==2022.6.15
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
Deprecated==1.2.13
distlib==0.3.6
Django==4.1.1
django-environ==0.9.0
filelock==3.8.0
identify==2.5.5
idna==3.3
jmespath==1.0.1
nodeenv==1.7.0
platformdirs==2.5.2
pre-commit==2.20.0
pycparser==2.21
PyGithub==1.55
PyJWT==2.4.0
PyNaCl==1.5.0
python-dateutil==2.8.2
PyYAML==6.0
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
urllib3==1.26.12
wrapt==1.14.1
xmltodict==0.13.0
